requirements.txt
fastapi
uvicorn
celery
redis
playwright
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
flower
httpx
logging
slowapi
google-genai